+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changes
 
+
+## Unreleased
+
+- Added two mechanisms to fix shape geometries based on config data.
+- Don't use cldfgeojson.dump since cutting off decimal places from coordinates may introduce
+  invalid geometries.
+
+
 ## [0.1.0] - 2024-11-22
 
 pyglottography provides